refactor: extract mock data to mocks folder, rename components, and extract ts types
