--- core/lobby/components/CreateGameRoom/GamePickModal/index.tsx
@@ -1,3 +1,0 @@
-import GamePickModal from "./GamePickModal";
-
-export default GamePickModal;
