+++ core/lobby/components/CreateRoomModal/GameListModal/GameList/CategoryTitle.tsx
