+++ core/lobby/components/CreateRoomModal/GameListModal/GameListModal.tsx
@@ -1,7 +1,6 @@
 import GameList from "./GameList";
-import styles from "./gamePickModal.module.css";
-import { GameType } from "../type";
-
+import styles from "./gameListModal.module.css";
+import { GameType } from "@/requests/games";
 interface GamePickModalProps {
   gameList: GameType[];
   activeGameId: string;
@@ -9,16 +8,16 @@ interface GamePickModalProps {
   onClose: () => void;
 }
 
-export default function GamePickModal({
+export default function GameListModal({
   onGameChange,
   onClose,
   gameList,
   activeGameId,
 }: GamePickModalProps) {
   return (
-    <div className="absolute left-0 w-full h-[80vh] lg:h-[500px] bg-[#292A2D] rounded-[10px] p-[25px]">
+    <div className="relative left-0 w-full h-[500px] bg-[#292A2D] rounded-[10px] py-[25px]">
       <button
-        className="absolute  w-[68px] h-[68px]  translate-y-[-100%] top-[-16px] lg:translate-x-[-100%] lg:top-[-6px] lg:left-[-36px] lg:translate-y-0  text-white rounded-[10px] bg-[#292A2D]"
+        className="absolute  w-[68px] h-[68px]  translate-y-[-100%] top-[-36px] lg:translate-x-[-100%] lg:top-[-30px] lg:left-[-48px] lg:translate-y-0  text-white rounded-[10px] bg-[#292A2D]"
         onClick={onClose}
       >
         {/* TODO replace inline-svg to Icon component */}
@@ -39,7 +38,7 @@ export default function GamePickModal({
           />
         </svg>
       </button>
-      <div className={`overflow-y-scroll h-full ${styles.scrollbar}`}>
+      <div className={`overflow-y-scroll h-full ${styles.scrollbar} pr-4`}>
         <GameList
           gameList={gameList}
           onGameChange={onGameChange}
