+++ core/lobby/components/CreateRoomModal/GameListModal/index.tsx
@@ -1,0 +1,3 @@
+import GameListModal from "./GameListModal";
+
+export default GameListModal;
