+++ core/lobby/components/CreateRoomModal/createRoomModall.module.css
@@ -1,5 +1,5 @@
 .radio__maxPlayers:checked ~ label {
-  box-shadow: 0 0 0 1px #2f88ff inset;
+  box-shadow: 0 0 0 2px #2f88ff inset;
 }
 
 .radio__maxPlayers:disabled ~ label {
