+++ pages/api/mock/rooms.ts
@@ -1,20 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { mock_createRoomResponse } from "@/mocks/rooms";
 
 export default function handler(req: NextApiRequest, res: NextApiResponse) {
-  return res.json({
-    id: "3345678",
-    name: "roomName",
-    game: {
-      id: "1",
-      name: "wolfKiller",
-    },
-    host: {
-      id: "abc",
-      nickname: "哈",
-    },
-    isLocked: false,
-    currentPlayers: 1,
-    minPlayers: 2,
-    maxPlayers: 6,
-  });
+  return res.json(mock_createRoomResponse);
 }
