+++ requests/games/index.ts
@@ -1,6 +1,15 @@
 import { IRequestWrapper, requestWrapper } from "@/requests/request";
 
-export const getAllGamesEndpoint = (): IRequestWrapper<{ url: string }> => {
+export type GameType = {
+  id: string;
+  displayName: string;
+  imageUrl: string;
+  minPlayers: number;
+  maxPlayers: number;
+  category?: string;
+};
+
+export const getAllGamesEndpoint = (): IRequestWrapper<GameType[]> => {
   return requestWrapper({
     url: "/api/internal/games",
     method: "GET",
