+++ requests/rooms/index.ts
@@ -1,16 +1,58 @@
 import { IRequestWrapper, requestWrapper } from "@/requests/request";
 
-export interface createRoomData {
+export type CreateRoomDataType = {
   name: string;
   gameId: string;
   password: number | null | "";
   minPlayers: number;
   maxPlayers: number;
+};
+
+export type CreateRoomResponseType = {
+  id: string;
+  name: string;
+  game: {
+    id: string;
+    name: string;
+  };
+  host: {
+    id: string;
+    nickname: string;
+  };
+  isLocked: boolean;
+  currentPlayers: number;
+  minPlayers: number;
+  maxPlayers: number;
+};
+
+export namespace Room {
+  export type Game = {
+    id: string;
+    name: string;
+  };
+
+  export type User = {
+    id: string;
+    nickname: string;
+    isReady: boolean;
+  };
+  export type Infomation = {
+    id: string;
+    name: string;
+    status: "WATTING" | "PLAYING";
+    game: Game;
+    host: User;
+    isLocked: boolean;
+    players: User[];
+    currentPlayers: number;
+    minPlayers: number;
+    maxPlayers: number;
+  };
 }
 
 export const createRoomEndpoint = (
-  data: createRoomData
-): IRequestWrapper<{ url: string }> => {
+  data: CreateRoomDataType
+): IRequestWrapper<CreateRoomResponseType> => {
   return requestWrapper({
     url: "/api/internal/rooms",
     method: "POST",
@@ -20,7 +62,7 @@ export const createRoomEndpoint = (
 
 export const getRoomInfoEndpoint = (
   roomId: string
-): IRequestWrapper<{ url: string }> => {
+): IRequestWrapper<Room.Infomation> => {
   return requestWrapper({
     url: `/api/internal/rooms/${roomId}`,
     method: "GET",
